+++ src/main/java/com/keylin/WeCare/dto/NannyDto.java
@@ -21,7 +21,7 @@ public class NannyDto {
 	@NotEmpty(message = "This field should not be empty")
 	private String postalCode;
 	@NotEmpty(message = "This field should not be empty")
-	private Float serviceFee;
+	private Double serviceFee;
 	@NotEmpty(message = "This field should not be empty")
 	private String description;
 	@NotEmpty(message = "This field should not be empty")
